Moved server side code out into systems directory.
